Fixes #3121 This can't have any side effect right?
git-svn-id: http://dev.piwik.org/svn/trunk@6448 59fd770c-687e-43c8-a1e3-f5a4ff64c105

+++ core/Db/Schema/Myisam.php
@@ -358,7 +358,7 @@ class Piwik_Db_Schema_Myisam implements Piwik_Db_Schema_Interface
 								  	  date2 DATE NULL,
 									  period TINYINT UNSIGNED NULL,
 								  	  ts_archived DATETIME NULL,
-								  	  value FLOAT NULL,
+								  	  value DOUBLE NULL,
 									  PRIMARY KEY(idarchive, name),
 									  INDEX index_idsite_dates_period(idsite, date1, date2, period, ts_archived),
 									  INDEX index_period_archived(period, ts_archived)
